Remove uniffi from user-facing binding names
All six language bindings now use clean, native-feeling names:
- Swift: `import Trafilatura` (module: Trafilatura, FFI: TrafilaturaFFI)
- Kotlin: `import trafilatura.*` (package: trafilatura)
- Ruby: `require "trafilatura"` (module: Trafilatura)
- Dart: `import 'trafilatura.dart'`
- C#: `using uniffi.trafilatura` (class: TrafilaturaMethods)
- JS: `import { Trafilatura } from './trafilatura.js'`

- Set `setup_scaffolding!("trafilatura")` for user-facing namespace
- Configure uniffi.toml with pretty names per language
- Pass --config to uniffi-bindgen for proper config discovery
- Change ExtractionConfig fields from i64 to i32 (fixes JS BigInt)
- Gitignore generated binding files (regenerated by make build-*)
- Update Makefile for new file names and paths
- Bump uniffi-bindgen-js to 0.2.1 (ffi_namespace fix)

+++ uniffi/src/exports.rs
@@ -73,10 +73,10 @@ pub enum HtmlDateMode {
 /// Advanced tuning parameters.
 #[derive(uniffi::Record)]
 pub struct ExtractionConfig {
-    pub min_extracted_size: i64,
-    pub min_extracted_comment_size: i64,
-    pub min_output_size: i64,
-    pub min_output_comment_size: i64,
+    pub min_extracted_size: i32,
+    pub min_extracted_comment_size: i32,
+    pub min_output_size: i32,
+    pub min_output_comment_size: i32,
 }
 
 /// Extraction options.
@@ -180,10 +180,10 @@ pub fn create_readable_document(result: ExtractResult) -> String {
 
 fn to_ffi_config(c: &trafilatura::Config) -> ExtractionConfig {
     ExtractionConfig {
-        min_extracted_size: c.min_extracted_size as i64,
-        min_extracted_comment_size: c.min_extracted_comment_size as i64,
-        min_output_size: c.min_output_size as i64,
-        min_output_comment_size: c.min_output_comment_size as i64,
+        min_extracted_size: c.min_extracted_size as i32,
+        min_extracted_comment_size: c.min_extracted_comment_size as i32,
+        min_output_size: c.min_output_size as i32,
+        min_output_comment_size: c.min_output_comment_size as i32,
     }
 }
